+++ Custom.txt
@@ -2,4 +2,24 @@
 => 벽을 넘어서 까지 체크 하는 경우가 있다.
 => 해결 방법	
 	시야각을 만들어 장애물을 체크 후 판단하기
+<<<<<<< Updated upstream
 	참고 : https://www.youtube.com/watch?v=rQG9aUWarwE
+=======
+	참고 : https://www.youtube.com/watch?v=rQG9aUWarwE
+
+추가 1 : 1인칭 이동 및 카메라 마우스 조절
+=> 카메라를 하나 추가하여 1인칭으로 볼수 있게 하였다.
+=> 문제점 1인칭이 되면 지붕이 없어 이상함 => 해결방법 ?: 지붕을 만들어 덮는다. 아직 생각
+					=> 기존 지붕을 지우고 Plane를 크기에 맞게 통으로 가려 사용했다.
+
+추가 예정 : 난이도 추가하기
+1 단계 유령이 없는 난이도
+2 단계 유령 추가
+3 단계 가고일 추가
+4 단계 3인칭 으로 변경
+5 단계 맵에 열쇠를 숨겨 찾아야 문이 열리는방법 
+
+열쇠 시스템 추가 필요
+
+추후 탈출을 위한 잠금장치추가 해보기
+>>>>>>> Stashed changes
