+++ RoomGame/Assets/MilkyWay.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 7d7573f338ffb584880826500aeb227b
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
