+++ RoomGame/Assets/MilkyWay/Material.meta
@@ -1,0 +1,9 @@
+fileFormatVersion: 2
+guid: 58d58d3b0df424d45aeb7d355d273b3e
+folderAsset: yes
+timeCreated: 1498423204
+licenseType: Store
+DefaultImporter:
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
