+++ RoomGame/Assets/MilkyWay/Material/MilkyWay.mat.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 860a66ca4259b384f9ec01b687f71c3c
+timeCreated: 1498417791
+licenseType: Store
+NativeFormatImporter:
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
