+++ RoomGame/Assets/MilkyWay/Texture.meta
@@ -1,0 +1,9 @@
+fileFormatVersion: 2
+guid: 011a5869475926f44b495bf414fe48f6
+folderAsset: yes
+timeCreated: 1498423211
+licenseType: Store
+DefaultImporter:
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
