+++ RoomGame/Assets/Scripts/Enemies/FindOfViewEditor.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: f85445b3549dc6a4e9e01b58b4362559
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
