--- RoomGame/Assets/Scripts/FindOfView.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: c476cac1b572efe4297241fc09ef4d61
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
