--- RoomGame/Assets/Scripts/FindOfViewEditor.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: f85445b3549dc6a4e9e01b58b4362559
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
