--- RoomGame/Assets/Scripts/FirstPersonView.cs
@@ -1,12 +1,0 @@
-using System.Collections;
-using System.Collections.Generic;
-using UnityEngine;
-
-public class FirstPersonView : MonoBehaviour
-{
-  
-   
-}
-
-   
-
