--- RoomGame/Assets/Scripts/FirstPersonView.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: fef7fcb1cd6e44c408a53aacfbe5a639
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
