--- RoomGame/Assets/Scripts/Observer.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: ff2d097703cef7d4784462d13b472583
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
