+++ RoomGame/Assets/Scripts/PlayerMovement.cs
@@ -4,22 +4,26 @@ using UnityEngine;
 
 public class PlayerMovement : MonoBehaviour
 {
-
     Rigidbody rigid_player;
     Animator animator_player;
     Vector3 movement;
-
-    public float turnSpeed;
+     
+    public float turnSpeed; //3인칭 이동속도
+  
     Quaternion rotation = Quaternion.identity;
 
     AudioSource AS_Foots;
 
-    public bool isKeyMove;
+    bool isKeyMove =true;
+
     //1인칭
-    public GameObject camera;
+    [Header("FirstPersonView")]
+    [SerializeField] GameObject camera;
+    [SerializeField] float moveSpeed; //1인칭 이동속도
+    [SerializeField] float rotSpeed = 200.0f;
     float mx,my; //마우스 이동 감지
     float camX, y; //이동 변수
-    public float rotSpeed = 200.0f;
+  
 
 
 
@@ -43,7 +47,7 @@ public class PlayerMovement : MonoBehaviour
             float _cameraRotationX = my * rotSpeed;
 
             camX -= _cameraRotationX;
-            //camX = Mathf.Clamp(camX, -cameraRotationLimit, cameraRotationLimit);
+            camX = Mathf.Clamp(camX, -20, 20);
 
             camera.transform.localEulerAngles = new Vector3(camX, 0f, 0f);
 
@@ -93,7 +97,7 @@ public class PlayerMovement : MonoBehaviour
         {
             Vector3 _moveHorizontal = transform.right * horizontal;
             Vector3 _moveVertical = transform.forward * vertical;
-            Vector3 _velocity = (_moveHorizontal + _moveVertical).normalized * turnSpeed;
+            Vector3 _velocity = (_moveHorizontal + _moveVertical).normalized * moveSpeed;
 
             rigid_player.MovePosition(transform.position + _velocity * Time.deltaTime);
         }
@@ -111,7 +115,10 @@ public class PlayerMovement : MonoBehaviour
         rigid_player.MoveRotation(rotation);
     }
 
-
+    public void ChangeMoveMode()
+    {
+        isKeyMove = !isKeyMove;
+    }
 
 
 }
