--- RoomGame/Assets/Scripts/WaypointPatrol.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 5a103036ef935234ea69ef50a31c431f
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
